+++ cmake/BLTPrivateMacros.cmake
@@ -1,4 +1,4 @@
-# Copyright (c) 2017-2022, Lawrence Livermore National Security, LLC and
+# Copyright (c) 2017-2023, Lawrence Livermore National Security, LLC and
 # other BLT Project Developers. See the top-level LICENSE file for details
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
@@ -628,11 +628,11 @@ endmacro(blt_make_file_ext_regex)
 ##
 ## Filters source list by file extension into C/C++, Fortran, Python, and
 ## CMake source lists based on BLT_C_FILE_EXTS, BLT_Fortran_FILE_EXTS,
-## and BLT_CMAKE_FILE_EXTS (global BLT variables). This filtering is
-## case-insensitive. Files named "CMakeLists.txt" are also filtered here.
-## Files with no extension or generator expressions that are not object
-## libraries (of the form "$<TARGET_OBJECTS:nameofobjectlibrary>") will
-## throw fatal errors.
+## BLT_Python_FILE_EXTS and BLT_CMAKE_FILE_EXTS (global BLT variables).
+## This filtering is case-insensitive. Files named "CMakeLists.txt" are
+## also filtered here. Files with no extension or generator expressions
+## that are not object libraries (of the form
+## "$<TARGET_OBJECTS:nameofobjectlibrary>") will throw fatal errors.
 ## ------------------------------------------------------------------------------
 macro(blt_split_source_list_by_language)
 
@@ -690,11 +690,11 @@ macro(blt_split_source_list_by_language)
             if (DEFINED arg_Fortran_LIST)
                 list(APPEND ${arg_Fortran_LIST} "${_file}")
             endif()
-        elseif("${_lower_file}" MATCHES "${BLT_Python_FILE_EXTS}")
+        elseif("${_lower_file}" MATCHES "${BLT_Python_FILE_REGEX}")
             if (DEFINED arg_Python_LIST)
                 list(APPEND ${arg_Python_LIST} "${_file}")
             endif()
-        elseif("${_lower_file}" MATCHES "${BLT_CMAKE_EXTS}" OR "${_name}" STREQUAL "CMakeLists.txt")
+        elseif("${_lower_file}" MATCHES "${BLT_CMAKE_REGEX}" OR "${_name}" STREQUAL "CMakeLists.txt")
             if (DEFINED arg_CMAKE_LIST)
                 list(APPEND ${arg_CMAKE_LIST} "${_file}")
             endif()
@@ -705,41 +705,6 @@ macro(blt_split_source_list_by_language)
     endforeach()
 
 endmacro(blt_split_source_list_by_language)
-
-
-##------------------------------------------------------------------------------
-## blt_update_project_sources( TARGET_SOURCES <sources> )
-##------------------------------------------------------------------------------
-macro(blt_update_project_sources)
-
-    set(options)
-    set(singleValueArgs)
-    set(multiValueArgs TARGET_SOURCES)
-
-    # Parse the arguments
-    cmake_parse_arguments(arg "${options}" "${singleValueArgs}"
-                            "${multiValueArgs}" ${ARGN} )
-
-    # Check arguments
-    if ( NOT DEFINED arg_TARGET_SOURCES )
-        message( FATAL_ERROR "Must provide target sources" )
-    endif()
-
-    ## append the target source to the all project sources
-    foreach( src ${arg_TARGET_SOURCES} )
-        if(IS_ABSOLUTE ${src})
-            list(APPEND "${PROJECT_NAME}_ALL_SOURCES" "${src}")
-        else()
-            list(APPEND "${PROJECT_NAME}_ALL_SOURCES"
-                "${CMAKE_CURRENT_SOURCE_DIR}/${src}")
-        endif()
-    endforeach()
-
-    set( "${PROJECT_NAME}_ALL_SOURCES" "${${PROJECT_NAME}_ALL_SOURCES}"
-        CACHE STRING "" FORCE )
-    mark_as_advanced("${PROJECT_NAME}_ALL_SOURCES")
-
-endmacro(blt_update_project_sources)
 
 
 ##------------------------------------------------------------------------------