+++ tests/smoke/blt_cuda_version_smoke.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2022, Lawrence Livermore National Security, LLC and
+// Copyright (c) 2017-2023, Lawrence Livermore National Security, LLC and
 // other BLT Project Developers. See the top-level LICENSE file for details
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
@@ -10,6 +10,7 @@
 //-----------------------------------------------------------------------------
 
 #include <iostream>
+#include <string>
 #include "cuda_runtime_api.h"