+++ tests/smoke/blt_gbenchmark_smoke.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2022, Lawrence Livermore National Security, LLC and
+// Copyright (c) 2017-2023, Lawrence Livermore National Security, LLC and
 // other BLT Project Developers. See the top-level LICENSE file for details
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
@@ -10,14 +10,14 @@
 #define BASIC_BENCHMARK_TEST(x) \
   BENCHMARK(x)->Arg( 1<<3 )->Arg( 1<<9 )->Arg( 1 << 13 )
 
-void benchmark_smoke_empty(benchmark::State& state)
+void benchmark_smoke_string_create(benchmark::State& state)
 {
   for (auto _ : state)
   {
-    benchmark::DoNotOptimize(state.iterations());
+    std::string empty_string;
   }
 }
-BENCHMARK(benchmark_smoke_empty);
+BENCHMARK(benchmark_smoke_string_create);
 
 
 void benchmark_smoke_spin_loop(benchmark::State& state)